Move legobridge to pkg/share/legobridge
