+++ pkg/certman2/controller/issuer/controlplane/acme/handler.go
@@ -19,9 +19,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	"github.com/gardener/cert-management/pkg/cert/legobridge"
 	"github.com/gardener/cert-management/pkg/certman2/apis/cert/v1alpha1"
 	"github.com/gardener/cert-management/pkg/certman2/core"
+	"github.com/gardener/cert-management/pkg/shared/legobridge"
 )
 
 // NewACMEIssuerHandler creates an ACME IssuerHandler.
