+++ pkg/controller/issuer/acme/handler.go
@@ -15,8 +15,8 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	api "github.com/gardener/cert-management/pkg/apis/cert/v1alpha1"
-	"github.com/gardener/cert-management/pkg/cert/legobridge"
 	"github.com/gardener/cert-management/pkg/controller/issuer/core"
+	"github.com/gardener/cert-management/pkg/shared/legobridge"
 )
 
 var acmeType = core.ACMEType
