+++ pkg/controller/issuer/revocation/reconciler.go
@@ -26,12 +26,12 @@ import (
 	"k8s.io/apimachinery/pkg/selection"
 
 	api "github.com/gardener/cert-management/pkg/apis/cert/v1alpha1"
-	"github.com/gardener/cert-management/pkg/cert/legobridge"
 	"github.com/gardener/cert-management/pkg/cert/source"
 	"github.com/gardener/cert-management/pkg/cert/utils"
 	ctrl "github.com/gardener/cert-management/pkg/controller"
 	"github.com/gardener/cert-management/pkg/controller/issuer/certificate"
 	"github.com/gardener/cert-management/pkg/controller/issuer/core"
+	"github.com/gardener/cert-management/pkg/shared/legobridge"
 )
 
 // RevokeReconciler creates a certificate revocation reconciler.
