+++ pkg/shared/legobridge/certificate_test.go
