+++ test/certman2/integration/controller/issuer/issuer_test.go
@@ -31,11 +31,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
 
-	"github.com/gardener/cert-management/pkg/cert/legobridge"
 	"github.com/gardener/cert-management/pkg/certman2/apis/cert/v1alpha1"
 	"github.com/gardener/cert-management/pkg/certman2/apis/config"
 	issuercontrolplane "github.com/gardener/cert-management/pkg/certman2/controller/issuer/controlplane"
 	"github.com/gardener/cert-management/pkg/certman2/core"
+	"github.com/gardener/cert-management/pkg/shared/legobridge"
 )
 
 var (
